discord_bot/requirements.txt
discord==1.0.1
python-dotenv==0.15.0
requests==2.25.1
tensorflow==2.4.1
pandas==1.2.4
numpy==1.20.2
pytorch==1.8.1
